+++ src/pytorch_lightning/trainer/trainer.py
@@ -2246,26 +2246,6 @@ class Trainer(
         self._predict_loop = loop
 
     @property
-    def verbose_evaluate(self) -> bool:
-        rank_zero_deprecation(
-            "The `Trainer.verbose_evaluate` property has been deprecated and will be removed in v1.8. The current value"
-            " returned is the union of the validate and test loop values. You can choose which one to access with"
-            " `trainer.{validate,test}_loop.verbose`.",
-            stacklevel=5,
-        )
-        return self.validate_loop.verbose or self.test_loop.verbose
-
-    @verbose_evaluate.setter
-    def verbose_evaluate(self, verbose: bool) -> None:
-        rank_zero_deprecation(
-            "The `Trainer.verbose_evaluate` property has been deprecated and will be removed in v1.8. This will set"
-            " the value for both trainer.{validate,test}_loop.verbose`.",
-            stacklevel=5,
-        )
-        self.validate_loop.verbose = verbose
-        self.test_loop.verbose = verbose
-
-    @property
     def _evaluation_loop(self) -> EvaluationLoop:
         if self.state.fn in (TrainerFn.FITTING, TrainerFn.TUNING):
             return self.fit_loop.epoch_loop.val_loop