+++ tests/tests_pytorch/deprecated_api/test_remove_1-8.py
@@ -72,15 +72,6 @@ def test_v1_8_0_deprecated_run_stage():
         trainer.run_stage()
 
 
-def test_v1_8_0_trainer_verbose_evaluate():
-    trainer = Trainer()
-    with pytest.deprecated_call(match="verbose_evaluate` property has been deprecated and will be removed in v1.8"):
-        assert trainer.verbose_evaluate
-
-    with pytest.deprecated_call(match="verbose_evaluate` property has been deprecated and will be removed in v1.8"):
-        trainer.verbose_evaluate = False
-
-
 @pytest.mark.parametrize("fn_prefix", ["validated", "tested", "predicted"])
 def test_v1_8_0_trainer_ckpt_path_attributes(fn_prefix: str):
     test_attr = f"{fn_prefix}_ckpt_path"